Breaking: Load Rusha uploader worker from NPM instead of JS file
Requires Rusha as a peer dependency if you are importing the ContentUploader as a React component. This is not required if you are using the vanilla JS wrapper. This simplifies our webpack logic, prevents Rusha from being loaded twice, and removes the need to special case for Rusha compared to other required third-party libraries.