Breaking: Render dropdowns and tooltips popups into body
Previously they used to render into the container element.
This has been causing issues when a UI Element was rendered
in some 3rd party library modal due to stacking context issues
and compatibility with tether.js positioning via bodyElement.

This change removes use of bodyElement making tether render
the popups into the docuemtn body. Now tether should be able to position better.

However the tooltip or dropdown menu can be invisibile or hidden
due to z-index issues. In that case, the parent app, folks using the
UI Element, need to override and add the z-index based on their app's
usecase. Classes to add z-index to would be `dropdown-menu-element` and
`tooltip-element`.