feat(i18n): use message-accumulator package to handle coded text
Coded text hides html and jsx components behind short xml-like tags
so that translators cannot muck with the contents of those html
or jsx tags. It also hides replacement parameters so that translations
from different i18n libraries and platforms can be shared with each
other. The contents of the html, jsx, or parameters are substituted
back in to the translated string before it is returned as a tree
of React components.