feat(badgeable): Pass down props to children
We have a pattern in EUA of using HeaderToggleButton to wrap a button and pass a
color prop to an icon depending on the state of the header and the theme. For
NotificationCenterButton and TaskFlyout we wrap the icons in a Badgeable as
well. Due to this, there's currently a bug where the the inner icon element is
not receiving the correct color prop from HeaderToggleButton, and this would fix
that issue.

BREAKING CHANGE: Children of the Badgeable component are restricted to
be a single React element