feat(i18n): Implement the new FormattedCompMessage component (#976)
BREAKING CHANGE: Added additional peer dependencies to support i18n FormattedCompMessage component

Implement a new component called `FormattedCompMessage`. This translates strings which can contain HTML or other JSX components in a safe way by hiding the contents of those tags from the translators or hackers and substituting the contents from source strings into the translations. Example:

```javascript
<FormattedCompMessage id="mystring" description="translator's comment">
   You can <span class="foo">delete</span> <Link to={url}>these files</Link>, or just
  <span id="x" class="unshare">unshare them</span>.
</FormattedCompMessage>
```

`FormattedCompMessage` is built as a layer on top of react-intl, and uses it to do the actual translations. Its function is to transform a tree of React elements into a coded string, get that string translated with react-intl, transform the coded translated string back into a tree of React elements, and finally substitute the hidden React elements from the source tree back into the translation tree in the right locations.

The component depends on the new babel plugin `babel-plugin-box-i18n` to extract the coded texts from the source code and place them into the same type of json files that the react-intl babel plugin produces. In this way, the translation process downstream from there works with no further changes. The translations appear in the properties files some time later and react-intl picks them up as normal. The `FormattedCompMessage` can retrieve these translations using react-intl. The `babel-pluging-box-i18n` plugin also enforces Box's policies, such as the requirement that all strings have a unique id and a description for the translators.

The `FormattedCompMessage` component can support plurals in a React-idiomatic way by embedding `<Plural>` components in the body of the `FormattedCompMessage`:

```javascript
<FormattedCompMessage id="mystring" description="translator's comment" count={count}>
  <Plural category="one">This is the <strong>singular</strong> string.</Plural>
  <Plural category="other">This is the <strong>plural</strong> string.</Plural>
</FormattedCompMessage>
```
Note that HTML and subcomponents are allowed in the plural string as well.

The `FormattedCompMessage` component supports replacement parameters in a React-idiomatic way by embedding `<Param>` components in the body of the `FormattedCompMessage`:

```javascript
<FormattedCompMessage id="mystring" description="translator's comment">
  User <Param value={user.name} description="The user's full name"/> has already deleted this file.
</FormattedCompMessage>
```

Note that the `Param` tags can be used inside of `Plural` tags as well, as they are normal components.