fix(uploads): Fix hanging on resuming uploads
* Fixes bug with subsequent uploadFile calls in Xhr.js failing
automatically after the original times out.
* Fixes hanging during logging of part failures by not blocking on it
* Fixes bug with listParts always returning undefined, causing the
client reupload a part it has already uploaded. Remove the byte check in
determining whether or not the file has been uploaded, because it has
shown to be unreliable.
* Removes PART_STATE_COMPUTING_DIGEST state as it is never used in the
code, and is not a useful state to keep track of.

Fixes most of WEBEN-1098, though I have still seen it hang from time to
time.