fix(folder-upload-node): fix subfolder name conflict errors
Omits incorrect setting of subfolder name conflict errors so that files
in subfolders can continue uploading after being added to the upload
queue.

If there are files/folders in error, the view would be set to
VIEW_ERROR and any subsequent uploads would be stopped until the upload
errors were resolved. In this case, subfolders will always hit a name
conflict so we shouldn't block uploading subfiles, and we want view to
be set to VIEW_UPLOAD_IN_PROGRESS to continue automatic uploads in
ContentUploader.updateViewAndCollection.